sframe.h: Add SFrame format definition
[Changes in V2]
  - Added an unsigned 8-bit integral field to SFrame header for
    supporting auxilliary SFrame header.  This helps make provisions to
    support any future ABI/arch by allowing each ABI/arch to add
    information specific to them in a flexible manner.
  - Changed the constants for
    SFRAME_FRE_TYPE_ADDR1/SFRAME_FRE_TYPE_ADDR2/SFRAME_FRE_TYPE_ADDR4
    from 1/2/3 to 0/1/2.
  - Other minor code formatting fixups.
[End of changes in V2]

The header sframe.h defines the SFrame format.

The SFrame format is the Simple Frame format.  It can be used to
represent the minimal necessary unwind information required for
backtracing.  The current version supports AMD64 and AARCH64.

More details of the SFrame format are included in the documentation
of the header file in this patch.

include/ChangeLog:
	* sframe.h: New file.