RISC-V: Optimize relax of GP/call with max_alignment.
The max_alignment defined out of [gp-2K, gp+2k), the max_alignment
shouldn't affect the relax of gp.
If the symbol is in [gp-2K, gp+2k), the max_alignment would be
replaced with the max_alignment of the section in [gp-2k, gp+2k).

Likewise, the max_alignment defined out of [call-1M, call+1M),
the max_alignment shouldn't affect the relax of call.
If the symbol is in [call-1M, call+1M), the max_alignment would be
replaced with the max_alignment of the section in [call-1M, call+1M).

bfd/
	* elfnn-riscv.c (_bfd_riscv_get_max_alignment_in_jtype): New.
	(_bfd_riscv_relax_call): The max_alignment of sections is from
	[base-1M, base+1M).
	(_bfd_riscv_get_max_alignment_in_itype): New.
	(_bfd_riscv_relax_lui): The max_alignment of sections is from
	[gp-2K, gp+2K).
	(_bfd_riscv_relax_pc): Likewise.
ld/
	* ld/testsuite/ld-riscv-elf/relax-max-align-call.*: New tests.
	* ld/testsuite/ld-riscv-elf/relax-max-align-gp.*: New tests.